ArmVirtPkg: enable AcpiTableDxe and EFI_ACPI_TABLE_PROTOCOL dynamically
In this patch, AcpiTableDxe, and the dependent (platform-matching) ACPI
table producing DXE drivers are built for 32-bit as well, and the ACPI
protocol / driver chain is enabled dynamically, when appropriate. This is
being done in one larger patch, because ArmVirt.dsc.inc, where
AcpiTableDxe is built, is used by all the platform DSCs.

No change in behavior should be observable after this patch on any
ArmVirtPkg platform.

Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Leif Lindholm <leif.lindholm@linaro.org>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>