OvmfPkg/PlatformPei: rebase to Q35TsegSizeLib
Replace

  FixedPcdGet8 (PcdQ35TsegMbytes)

macro invocations with

  Q35TsegSizeGetPreferredMbytes ()

function invocations. This causes no change in observable behavior.

Remember that Q35TsegSizeGetPreferredMbytes() -- indirectly -- logs an
error and calls CpuDeadLoop() if it is executed on a non-Q35 board.  This
is safe and intentional, because all the converted call sites are reached
only when the PcdSmmSmramRequire Feature PCD is set, for which Q35 has
always been a hard requirement.

Cc: Jordan Justen <jordan.l.justen@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>