OvmfPkg/IoMmuDxe: shut up "unused-const-variable" gcc-6 warning in RELEASE
Starting with gcc-6, a new warning option called "-Wunused-const-variable"
has become available (and enabled by default under our build settings):

  https://gcc.gnu.org/onlinedocs/gcc-6.4.0/gcc/Warning-Options.html

We should give it the same treatment as Ard gave "unused-but-set-variable"
in commit 20d00edf21d2 ("BaseTools/GCC: set -Wno-unused-but-set-variables
only on RELEASE builds", 2016-03-24); i.e., we should restrict the warning
to the DEBUG build target.

However, because the new warning is gcc-6+ only, we cannot add
"-Wno-unused-const-variable" to any GCC5 macros in
"BaseTools/Conf/tools_def.template". While the GCC6 toolchain and/or the
desired handling of the new warning are investigated in
<https://bugzilla.tianocore.org/show_bug.cgi?id=700>, suppress the warning
for gcc-6+ (in RELEASE builds) as follows:

- Replace the "mBusMasterOperationName" array with the
  BUS_MASTER_OPERATION_NAME(Expression, ShortName) macro that compares
  Expression against EdkiiIoMmuOperationBusMaster<ShortName>, and in case
  of a match, evaluates to "ShortName", stringified,

- when composing the DEBUG message -- which the preprocessor might
  eliminate from RELEASE builds, thereby removing its references to
  variables --, build a ladder of comparisons with
  BUS_MASTER_OPERATION_NAME().

Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Brijesh Singh <brijesh.singh@amd.com>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Thomas Lamprecht <t.lamprecht@proxmox.com>
Reported-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Laszlo Ersek <lersek@redhat.com>